Progress Thus Far
I followed along and completed the task as describe in the first three vidoes of the following YouTube playlist: https://www.youtube.com/watch?v=gKzSCf0VdCQ&list=PL3hpmQhMoz-cz1GBAtovJyrfspZctG03L&index=2.  While doing so, I made adjustments to utilize a username instead of an email; to include a security question and answer in the webforms, database, and user class; and to eliminate the concept of authentication roles since it was beyond the scope of the requested project.